+++ lib/src/internal/qa/table_area_layout_qa.dart
@@ -92,6 +92,7 @@ class _HomePageState extends State<HomePage> {
         setHoveredRowIndex: _setHoveredRowIndex,
         hoveredRowIndex: 0,
         onRowTap: null,
+        onRowSecondaryTap: null,
         onRowDoubleTap: null,
         scrollBehavior: scrollBehavior);