+++ lib/src/internal/qa/table_row_qa.dart
@@ -65,6 +65,7 @@ class _HomePageState extends State<HomePage> {
         setHoveredRowIndex: _setHoveredRowIndex,
         hoveredRowIndex: 0,
         onRowTap: null,
+        onRowSecondaryTap: null,
         onRowDoubleTap: null);
 
     return Scaffold(